// src/app/core/components/admin/boat-model/models/boat-model.ts
export class BoatModel{
    _id: number;
    manufacturerId: number;
    manufacturerName: string;
    modelName: string;
    typeId: number;
    typeName: string;
    lengthMeter: number;
    beamMeter: number;
    draftMeter: number;
    boatHullMetarialId: number;
    boatHullMetarialName: string;
    isDeleted: boolean;
    createdAt: Date | any;
    updatedAt: Date | any;
  }
  